use uint16 for fee

+++ uniswap.hpp
@@ -13,7 +13,7 @@ namespace uniswap {
      * - `{uint64_t} amount_in` - amount input
      * - `{uint64_t} reserve_in` - reserve input
      * - `{uint64_t} reserve_out` - reserve output
-     * - `{uint8_t} [fee=30]` - (optional) trade fee (pips 1/100 of 1%)
+     * - `{uint16_t} [fee=30]` - (optional) trade fee (pips 1/100 of 1%)
      *
      * ### example
      *
@@ -22,14 +22,14 @@ namespace uniswap {
      * const uint64_t amount_in = 10000;
      * const uint64_t reserve_in = 45851931234;
      * const uint64_t reserve_out = 125682033533;
-     * const uint8_t fee = 30;
+     * const uint16_t fee = 30;
      *
      * // Calculation
      * const uint64_t amount_out = uniswap::get_amount_out( amount_in, reserve_in, reserve_out, fee );
      * // => 27328
      * ```
      */
-    static uint64_t get_amount_out( const uint64_t amount_in, const uint64_t reserve_in, const uint64_t reserve_out, const uint8_t fee = 30 )
+    static uint64_t get_amount_out( const uint64_t amount_in, const uint64_t reserve_in, const uint64_t reserve_out, const uint16_t fee = 30 )
     {
         // checks
         eosio::check(amount_in > 0, "SX.Uniswap: INSUFFICIENT_INPUT_AMOUNT");
@@ -54,7 +54,7 @@ namespace uniswap {
      * - `{uint64_t} amount_out` - amount input
      * - `{uint64_t} reserve_in` - reserve input
      * - `{uint64_t} reserveOut` - reserve output
-     * - `{uint8_t} [fee=30]` - (optional) trading fee (pips 1/100 of 1%)
+     * - `{uint16_t} [fee=30]` - (optional) trading fee (pips 1/100 of 1%)
      *
      * ### example
      *
@@ -63,14 +63,14 @@ namespace uniswap {
      * const uint64_t amount_out = 27328;
      * const uint64_t reserve_in = 45851931234;
      * const uint64_t reserve_out = 125682033533;
-     * const uint8_t fee = 30;
+     * const uint16_t fee = 30;
      *
      * // Calculation
      * const uint64_t amount_in = uniswap::get_amount_in( amount_out, reserve_in, reserve_out, fee );
      * // => 10000
      * ```
      */
-    static uint64_t get_amount_in( const uint64_t amount_out, const uint64_t reserve_in, const uint64_t reserve_out, const uint8_t fee = 30 )
+    static uint64_t get_amount_in( const uint64_t amount_out, const uint64_t reserve_in, const uint64_t reserve_out, const uint16_t fee = 30 )
     {
         // checks
         eosio::check(amount_out > 0, "SX.Uniswap: INSUFFICIENT_OUTPUT_AMOUNT");
